+++ python-code/src/amr_hub_abm/task/task.py
@@ -26,6 +26,7 @@ from dataclasses import dataclass, field
 from enum import IntEnum
 from typing import TYPE_CHECKING
 
+from amr_hub_abm.agent.utils import add_agent_occupancy, remove_agent_occupancy
 from amr_hub_abm.exceptions import SimulationModeError, TimeError
 from amr_hub_abm.space.location import Location
 from amr_hub_abm.space.space import check_if_location_reached, get_room
@@ -38,64 +39,6 @@ if TYPE_CHECKING:
 
 
 logger = logging.getLogger(__name__)
-
-
-# ---------------------------------------------------------------------------
-# Occupancy helpers
-#
-# NOTE: these mutate Agent + Content state and arguably belong on Agent or
-# Content (e.g. ``content.occupy(agent)`` / ``content.release()``). They are
-# kept here for now so this refactor only touches task.py; move them when the
-# Agent refactor lands.
-# ---------------------------------------------------------------------------
-
-
-def remove_agent_occupancy(agent: Agent, current_time: int) -> None:
-    """
-    Remove the agent's occupancy from any content they are currently occupying.
-
-    Parameters
-    ----------
-    agent : Agent
-        The agent whose occupancy is to be removed.
-    current_time : int
-        The current time in the simulation, used for logging purposes.
-
-    """
-    room = get_room(agent.location, agent.rooms)
-    if room is None:
-        return
-    for content in room.contents:
-        if content.occupier_id == (agent.idx, agent.agent_type):
-            content.occupier_id = None
-            agent.stationary = False
-            logger.info(
-                "Agent %s released content %s (%s) in room %s at t=%d.",
-                agent.idx,
-                content.content_id,
-                content.content_type,
-                room.name,
-                current_time,
-            )
-            return
-
-
-def add_agent_occupancy(agent: Agent, content: Content, current_time: int) -> None:
-    """Mark the agent as occupying the given content."""
-    content.occupier_id = (agent.idx, agent.agent_type)
-    agent.stationary = True
-
-    room = get_room(agent.location, agent.rooms)
-
-    logger.info(
-        "Agent %s occupied content %s (%s) in room %s at t=%d.",
-        agent.idx,
-        content.content_id,
-        content.content_type,
-        "unknown" if room is None else room.name,
-        current_time,
-    )
-
 
 # ---------------------------------------------------------------------------
 # Enums
